Return nearest location when there are multiple candidates
