+++ lib/dialogs/entity_lookup.js
@@ -10,6 +10,9 @@
 // See COPYING for details
 "use strict";
 
+const Ast = require('thingtalk').Ast;
+const { tryGetCurrentLocation, distance } = require('../utils');
+
 function tokensEquals(one, two) {
     if (one === two)
         return true;
@@ -83,11 +86,20 @@ async function lookupEntity(dlg, entityType, entityDisplay) {
 async function lookupLocation(dlg, searchKey) {
     let candidates = await dlg.manager.thingpedia.lookupLocation(searchKey);
     // ignore locations larger than a city
-    candidates = candidates.filter((c) => c.rank <= 16);
+    candidates = candidates.filter((c) => c.rank <= 16).map((c) => {
+        return new Ast.Location.Absolute(c.latitude, c.longitude, c.display);
+    });
 
     if (candidates.length === 0) {
         await dlg.reply(dlg._("Sorry, I cannot find any location matching “%s”.").format(searchKey));
         return null;
+    }
+
+    const currentLocation = await tryGetCurrentLocation(dlg);
+    if (currentLocation) {
+        candidates.sort((l1, l2) => {
+            return distance(l1, currentLocation.value) - distance(l2, currentLocation.value);
+        });
     }
 
     return candidates[0];
