+++ lib/dialogs/slot_filling.js
@@ -190,7 +190,7 @@ async function concretizeValue(dlg, slot) {
         const resolved = await lookupLocation(dlg, value.value.name);
         if (resolved === null)
             return false;
-        value.value = new Ast.Location.Absolute(resolved.latitude, resolved.longitude, resolved.display);
+        value.value = resolved;
     } else if (value.isLocation && value.value.isRelative) {
         const resolved = await resolveUserContext(dlg, '$context.location.' + value.value.relativeTag);
         if (resolved)
