+++ lib/dialogs/user_context.js
@@ -14,19 +14,7 @@ const Ast = ThingTalk.Ast;
 
 const ValueCategory = require('../semantic').ValueCategory;
 const { lookupLocation } = require('./entity_lookup');
-
-async function tryGetCurrentLocation(dlg) {
-    const gps = dlg.manager.platform.getCapability('gps');
-    if (gps === null)
-        return null;
-    const location = await gps.getCurrentLocation();
-    if (location === null) {
-        console.log('GPS location not available');
-        return null;
-    } else {
-        return new Ast.Value.Location(new Ast.Location.Absolute(location.latitude, location.longitude, location.display||null));
-    }
-}
+const { tryGetCurrentLocation } = require('../utils');
 
 function valueFromJs(category, value) {
     if (category.isLocation)
@@ -106,7 +94,7 @@ module.exports = async function resolveUserContext(dlg, variable) {
             const resolved = await lookupLocation(dlg, answer.value.name);
             if (resolved === null)
                 return false;
-            answer = new Ast.Value.Location(new Ast.Location.Absolute(resolved.latitude, resolved.longitude, resolved.display));
+            answer = new Ast.Value.Location(resolved);
         }
         sharedPrefs.set('context-' + variable, answer.toJS());
     }
